requirements.txt
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
cattrs==1.4.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
dateparser==1.0.0
distlib==0.3.1
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
lxml==4.6.3
MarkupSafe==1.1.1
oauthlib==3.1.0
parse==1.19.0
Pillow>=8.3.2
numpy==1.20.1
regex==2021.4.4
requests==2.25.1
requests-html==0.10.0
